This commit provides general fixes for 64-bit integer conversions woes, function reference problems with push / set, boolean conversion warnings in VC++, and a few other issues that prevented compilation with VS 2015.2

+++ lib/luwra/methods.hpp
@@ -38,7 +38,7 @@ namespace internal {
 
 		template <MethodPointerType meth> static inline
 		int invoke(State* state) {
-			return map<FunctionSignature>(state, hook<meth>);
+			return static_cast<int>(map<FunctionSignature>(state, hook<meth>));
 		}
 	};
 
@@ -55,7 +55,7 @@ namespace internal {
 
 		template <MethodPointerType meth> static inline
 		int invoke(State* state) {
-			return map<FunctionSignature>(state, hook<meth>);
+			return static_cast<int>(map<FunctionSignature>(state, hook<meth>));
 		}
 	};
 
@@ -89,7 +89,7 @@ namespace internal {
 
 		template <MethodPointerType meth> static inline
 		int invoke(State* state) {
-			return map<FunctionSignature>(state, hook<meth>);
+			return static_cast<int>(map<FunctionSignature>(state, hook<meth>));
 		}
 	};
 }
